+++ applications/smurf/libsmf/smf_rebincube_seqf.c
@@ -23,7 +23,7 @@
 *                              int ndim_out, const int lbnd_out[],
 *                              const int ubnd_out[], const int lbnd[],
 *                              const int ubnd[], float out[], float out_var[],
-*                              double weights[], int *nused, int *status );
+*                              double weights[], size_t *nused, int *status );
 
 *  Arguments:
 *     workforce = ThrWorkForce * (Given)
@@ -122,7 +122,7 @@ void smf_rebincube_seqf( ThrWorkForce *workforce, int njobs,
                          int ndim_out, const int lbnd_out[],
                          const int ubnd_out[], const int lbnd[],
                          const int ubnd[], float out[], float out_var[],
-                         double weights[], int *nused, int *status ){
+                         double weights[], size_t *nused, int *status ){
 
 /* Local Variables */
    int i;                            /* Thread index */
